My Code Review Changes
